Fix DB schema type mismatch, setup-db.php SQL parser, and test flakiness
- email_queue.campaign_id and template_id were INT but referenced BIGINT
  primary keys, causing "errno 150" FK constraint failures on MariaDB
- setup-db.php's statement splitter discarded any SQL statement preceded
  by a comment line in the same chunk, silently skipping most of the
  schema; now strips comments before splitting, and no longer wraps DDL
  in an explicit transaction (CREATE TABLE auto-commits in MySQL/MariaDB,
  which broke the trailing commit() call)
- QueueWorkerTest: COUNT(*) returns a PHP int under native prepares, not
  a string - loosened an overly strict assertSame comparison
- Verified end-to-end against the real ALL-INKL production database:
  schema installs cleanly (18 tables), all Unit and Integration tests
  pass, admin login and dashboard render correctly on the live server
- composer.json: moved phpstan to require-dev (was incorrectly forcing
  a static analysis tool into production installs)
- Added composer.lock for reproducible installs

+++ database/schema.sql
@@ -216,13 +216,13 @@ CREATE TABLE IF NOT EXISTS campaign_steps (
 -- ----------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS email_queue (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    campaign_id INT NOT NULL,
+    campaign_id BIGINT NOT NULL,
     contact_id BIGINT NULL,
     recipient_email VARCHAR(255) NOT NULL,
     subject VARCHAR(255) NOT NULL,
     body_html LONGTEXT,
     body_text TEXT,
-    template_id INT,
+    template_id BIGINT,
     sender_email VARCHAR(255),
 
     -- Status Tracking
